[ruby] Remove dummy mutation
As a newbie to GraphQL this part of the tutorial confuses me: why would I declare a mutation that has a "getter" name (`allLinks`) and does not mutate anything?

From the history of the tuto I figured that the graphql-ruby `install` generator used to create an empty Mutation type and this code was just a filler to prevent an error. It's not the case anymore though: a dummy "testMutation" is created by default by the generator.

So I think it's less confusing to remove this part from the tutorial.